Give application access to primary DO stub
We give a replica DO access to its primary through a DurableObject stub.

This is a pretty big departure from our typical stub creation flow.
Usually we need to make an ActorChannel with an outgoing factory, which
needs to make an outgoing request (internally) to get a capability to
the receiving Durable Object.

In this case, we already have our capability to the primary Durable
Object once the ActorCacheInterface is up and running. Therefore,
all we need to do is create an ActorChannel and give it access to the
ActorCacheInterface, which will take care of the rest (internally).

Co-authored-by: Justin Mazzola Paluska <jmp@cloudflare.com>